test: cover the mandatory DC value guards in the battery read loop
Adds tests for the non-CAN V/I/P guard and the temperature guard:

- a None voltage, current, power or temperature each fail with a
  ValueError naming the battery and the values, not a bare TypeError
- the read trial is still counted and nothing is published from a failed
  read, and past READ_TRIALS the driver still exits for a restart: the
  guards make the failure diagnosable, they do not prevent it
- no partial accumulation happens before the raise, asserted against the
  running totals in the frame locals the read-failure handler dumps
- with all values present the aggregate is unchanged (voltage and
  temperature averaged over NR_OF_BATTERIES, current and power summed)
- 0 A / 0 W / 0 degC are not mistaken for missing values

The scaffolding that loads the driver off-device (sys.modules stubs for
dbus/vedbus/gi, load by path, scripted D-Bus monitor, recording fake of
the gated D-Bus service) moves from the extrema test module into
tests/driver_harness.py so both test modules share it. No test logic
changed in the move.
